Modified retete_medicamente table

+++ pharmacy-management-system/src/main/java/com/bdproject/pharmacy/repository/RecipeMedicineRepository.java
@@ -1,6 +1,5 @@
 package com.bdproject.pharmacy.repository;
 
-import com.bdproject.pharmacy.model.MedicineEntity;
 import com.bdproject.pharmacy.model.RecipeEntity;
 import com.bdproject.pharmacy.model.RecipeMedicineEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -9,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface RecipeMedicineRepository extends JpaRepository<RecipeMedicineEntity, Long> {
+public interface RecipeMedicineRepository extends JpaRepository<RecipeMedicineEntity, Integer> {
 
     List<RecipeMedicineEntity> findByReteta(RecipeEntity recipe);
 }
